refactor(playerprofile): Update package name from `player` to `playerprofile`
