+++ backend/src/main/java/com/project/G1_T3/leaderboard/service/LeaderboardService.java
@@ -2,11 +2,12 @@ package com.project.G1_T3.leaderboard.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import com.project.G1_T3.player.repository.PlayerProfileRepository;
+
 import com.project.G1_T3.user.model.User;
 import com.project.G1_T3.user.service.UserService;
 import com.project.G1_T3.leaderboard.model.LeaderboardPlayerProfile;
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.repository.PlayerProfileRepository;
 
 import java.util.List;
 import java.util.Optional;
