+++ backend/src/main/java/com/project/G1_T3/match/service/MatchServiceImpl.java
@@ -5,11 +5,11 @@ import com.project.G1_T3.common.model.Status;
 import com.project.G1_T3.match.model.Match;
 import com.project.G1_T3.match.model.MatchDTO;
 import com.project.G1_T3.match.repository.MatchRepository;
-import com.project.G1_T3.player.model.PlayerProfile;
-import com.project.G1_T3.player.repository.PlayerProfileRepository;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.repository.PlayerProfileRepository;
+import com.project.G1_T3.playerprofile.service.PlayerProfileService;
 
 import lombok.extern.slf4j.Slf4j;
-import com.project.G1_T3.player.service.PlayerProfileService;
 
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.http.HttpStatus;
