+++ backend/src/main/java/com/project/G1_T3/matchmaking/service/MatchmakingService.java
@@ -2,7 +2,8 @@ package com.project.G1_T3.matchmaking.service;
 
 import com.project.G1_T3.match.model.Match;
 import com.project.G1_T3.matchmaking.model.MatchNotification;
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+
 import java.util.UUID;
 
 public interface MatchmakingService {
