+++ backend/src/main/java/com/project/G1_T3/matchmaking/service/PlayerQueue.java
@@ -1,6 +1,6 @@
 package com.project.G1_T3.matchmaking.service;
 
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
 
 public interface PlayerQueue {
     void addPlayer(PlayerProfile player, double latitude, double longitude);
