+++ backend/src/main/java/com/project/G1_T3/playerprofile/repository/PlayerProfileRepository.java
@@ -1,10 +1,12 @@
-package com.project.G1_T3.player.repository;
+package com.project.G1_T3.playerprofile.repository;
 
-import com.project.G1_T3.player.model.PlayerProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
+
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+
 import java.util.List;
 import java.util.UUID;
