+++ backend/src/main/java/com/project/G1_T3/round/model/RoundDTO.java
@@ -2,7 +2,7 @@ package com.project.G1_T3.round.model;
 
 import com.project.G1_T3.common.model.Status;
 import com.project.G1_T3.match.model.Match;
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
 import com.project.G1_T3.stage.model.Format;
 
 import lombok.Getter;
