+++ backend/src/main/java/com/project/G1_T3/stage/service/StageServiceImpl.java
@@ -1,7 +1,7 @@
 package com.project.G1_T3.stage.service;
 
 import com.project.G1_T3.common.model.Status;
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
 import com.project.G1_T3.round.service.RoundService;
 import com.project.G1_T3.stage.model.Format;
 import com.project.G1_T3.stage.model.Stage;
