+++ backend/src/main/java/com/project/G1_T3/tournament/controller/TournamentController.java
@@ -1,6 +1,6 @@
 package com.project.G1_T3.tournament.controller;
 
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
 import com.project.G1_T3.tournament.model.Tournament;
 import com.project.G1_T3.tournament.model.TournamentDTO;
 import com.project.G1_T3.tournament.service.TournamentService;
