+++ backend/src/test/java/com/project/G1_T3/match/MatchServiceTest.java
@@ -3,8 +3,8 @@ package com.project.G1_T3.match;
 import com.project.G1_T3.match.model.Match;
 import com.project.G1_T3.match.model.MatchDTO;
 import com.project.G1_T3.match.service.MatchServiceImpl;
-import com.project.G1_T3.player.model.PlayerProfile;
-import com.project.G1_T3.player.service.PlayerProfileService;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.service.PlayerProfileService;
 import com.project.G1_T3.match.repository.MatchRepository;
 import com.project.G1_T3.common.model.Status;
