+++ backend/src/test/java/com/project/G1_T3/matchmaking/service/EloMatchmakingAlgorithmTests.java
@@ -3,7 +3,7 @@ package com.project.G1_T3.matchmaking.service;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import com.project.G1_T3.matchmaking.model.QueuedPlayer;
-import com.project.G1_T3.player.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
 
 import static org.junit.jupiter.api.Assertions.*;
