+++ backend/src/test/java/com/project/G1_T3/round/RoundServiceTest.java
@@ -3,8 +3,8 @@ package com.project.G1_T3.round;
 import com.project.G1_T3.match.model.Match;
 import com.project.G1_T3.match.model.MatchDTO;
 import com.project.G1_T3.match.service.MatchService;
-import com.project.G1_T3.player.model.PlayerProfile;
-import com.project.G1_T3.player.repository.PlayerProfileRepository;
+import com.project.G1_T3.playerprofile.model.PlayerProfile;
+import com.project.G1_T3.playerprofile.repository.PlayerProfileRepository;
 import com.project.G1_T3.round.model.Round;
 import com.project.G1_T3.round.repository.RoundRepository;
 import com.project.G1_T3.round.service.RoundServiceImpl;
